Update Card.java
Had to add some new private variables because of some weird conflict because of a static context I didn't really understand it but I still changed the program.  I created the Value variable to represent the value of the card and made appropriate methods.  Did the same by adding a new private variable Suit which is a string of the suit name as well as appropriate methods to set and get that value.